latest

+++ jaxrs-sample/src/main/java/com/andre/pphere/dao/OfferDao.java
@@ -8,8 +8,8 @@ import java.util.*;
  */
 public interface OfferDao {
 	public List<Offer> getOffers(Double lat, Double lng, Double radius) ;
-	public Offer getOffer(int id) ;
-	public int createOffer(Offer store);
+	public Offer getOffer(long id) ;
+	public long createOffer(Offer store);
 	public void updateOffer(Offer store);
-	public void deleteOffer(int id);
+	public void deleteOffer(long id);
 }
