+++ jaxrs-sample/src/main/java/com/andre/pphere/data/Offer.java
@@ -17,7 +17,7 @@ import javax.xml.bind.annotation.XmlType;
 public class Offer extends BaseObject {
 	public Offer() {} 
 
-	public Offer(Integer id) {
+	public Offer(Long id) {
 		setId(id);
 	}
